requirements.txt
# Base requirements
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
drf-yasg>=1.21,<2.0
Pillow>=10.1,<11.0
python-dotenv>=1.0,<2.0
whitenoise>=6.6,<7.0
djoser 
stripe 
python-decouple

# Database
psycopg2-binary>=2.9,<3.0
pgvector 

# Redis & Celery
redis>=5.0,<6.0
django-redis>=5.4,<6.0
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Production
gunicorn>=21.2,<22.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code quality
flake8>=6.1,<7.0
black>=23.12,<24.0
isort>=5.13,<6.0